move editor test code
Moves the following functions which are only used for testing in the
test folder.

* editor_buffer_get_lastchar
* editor_buffer_get_cursor
* editor_buffer_set_cursor
* editor_buffer_set
