+++ test/editor/common.h
@@ -1,9 +1,9 @@
 /**
  * @file
- * Test code for editor_buffer_get_lastchar()
+ * Shared Editor Buffer Testing Code
  *
  * @authors
- * Copyright (C) 2023 Richard Russon <rich@flatcap.org>
+ * Copyright (C) 2022 Richard Russon <rich@flatcap.org>
  *
  * @copyright
  * This program is free software: you can redistribute it and/or modify it under
@@ -20,33 +20,15 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#define TEST_NO_MAIN
-#include "config.h"
-#include "acutest.h"
+#ifndef TEST_EDITOR_COMMON_H
+#define TEST_EDITOR_COMMON_H
+
 #include <stddef.h>
-#include <stdbool.h>
-#include "email/lib.h"
-#include "core/lib.h"
 #include "editor/lib.h"
 
-void test_editor_buffer_get_lastchar(void)
-{
-  // size_t editor_buffer_get_lastchar(struct EnterState *es);
+size_t editor_buffer_get_lastchar(struct EnterState *es);
+size_t editor_buffer_get_cursor(struct EnterState *es);
+size_t editor_buffer_set_cursor(struct EnterState *es, size_t pos);
+int editor_buffer_set(struct EnterState *es, const char *str);
 
-  {
-    TEST_CHECK(editor_buffer_get_lastchar(NULL) == 0);
-  }
-
-  {
-    struct EnterState *es = enter_state_new();
-    TEST_CHECK(editor_buffer_get_lastchar(es) == 0);
-    enter_state_free(&es);
-  }
-
-  {
-    struct EnterState *es = enter_state_new();
-    editor_buffer_set(es, "hello world");
-    TEST_CHECK(editor_buffer_get_lastchar(es) == 11);
-    enter_state_free(&es);
-  }
-}
+#endif /* TEST_EDITOR_COMMON_H */
