+++ test/editor/editor_kill_whole_line.c
@@ -22,6 +22,7 @@
 
 #define TEST_NO_MAIN
 #include "config.h"
+#include "common.h"
 #include "acutest.h"
 #include <stddef.h>
 #include <stdbool.h>
